fix(backup): widen IBigFileGate.LargeFileThresholdBytes to long and document KB conversion
Two follow-up tweaks on the IBigFileGate interface:

- Type was int but fileSizeBytes (and FileInfo.Length, LogEntry.FileSize)
  is long, so every comparison silently up-cast the threshold. Use long
  for parity and to allow thresholds beyond 2 GB.
- AppSettings stores the threshold as LargeFileThresholdKb (kilobytes,
  default 4096). The interface exposes bytes for ergonomic comparison
  with fileSizeBytes. Spell out in the comment that the concrete
  implementation converts via *1024 so future implementors don't trip
  on the unit mismatch.

+++ src/EasySave/Services/IBigFileGate.cs
@@ -18,8 +18,13 @@ namespace EasySave.Services;
 public interface IBigFileGate
 {
     // Threshold above which (>=) a file is considered "large" and gated.
-    // Loaded from settings by the concrete implementation.
-    int LargeFileThresholdBytes { get; }
+    // Expressed in bytes for direct comparison with fileSizeBytes; the
+    // concrete implementation is responsible for reading AppSettings
+    // .LargeFileThresholdKb and multiplying by 1024 to land here. long
+    // matches the parameter type below and the rest of the codebase
+    // (FileInfo.Length, LogEntry.FileSize) so the comparison never loses
+    // bits on multi-GB thresholds.
+    long LargeFileThresholdBytes { get; }
 
     // Acquires a slot for the file about to be transferred. Returns
     // immediately if fileSizeBytes < LargeFileThresholdBytes (small file —
